Move shared `compile_*` to `BottomUpBuilder`
Now, the SDD builder gets quite a bit of behaviour for free!

(also, renames `from_boolexpr` to `from_logical_expr` for
consistency)
