+++ src/builder/bdd/builder.rs
@@ -1,11 +1,9 @@
 use crate::{
     builder::{bdd::CompiledCNF, BottomUpBuilder},
-    plan::BottomUpPlan,
     repr::{
         bdd::{BddNode, BddPtr},
         cnf::Cnf,
         ddnnf::DDNNFPtr,
-        logical_expr::LogicalExpr,
         model::PartialModel,
         var_label::VarLabel,
     },
@@ -81,66 +79,6 @@ pub trait BddBuilder<'a>: BottomUpBuilder<'a, BddPtr<'a>> {
         ptr
     }
 
-    /// Compile a BDD from a CNF
-    fn compile_cnf(&'a self, cnf: &Cnf) -> BddPtr<'a> {
-        let mut cvec: Vec<BddPtr> = Vec::with_capacity(cnf.clauses().len());
-        if cnf.clauses().is_empty() {
-            return BddPtr::true_ptr();
-        }
-        // check if there is an empty clause -- if so, UNSAT
-        if cnf.clauses().iter().any(|x| x.is_empty()) {
-            return BddPtr::false_ptr();
-        }
-
-        // sort the clauses based on a best-effort bottom-up ordering of clauses
-        let mut cnf_sorted = cnf.clauses().to_vec();
-        cnf_sorted.sort_by(|c1, c2| {
-            // order the clause with the first-most variable last
-            let fst1 = c1
-                .iter()
-                .max_by(|l1, l2| {
-                    if self.less_than(l1.label(), l2.label()) {
-                        Ordering::Less
-                    } else {
-                        Ordering::Equal
-                    }
-                })
-                .unwrap();
-            let fst2 = c2
-                .iter()
-                .max_by(|l1, l2| {
-                    if self.less_than(l1.label(), l2.label()) {
-                        Ordering::Less
-                    } else {
-                        Ordering::Equal
-                    }
-                })
-                .unwrap();
-            if self.less_than(fst1.label(), fst2.label()) {
-                Ordering::Less
-            } else {
-                Ordering::Equal
-            }
-        });
-
-        for lit_vec in cnf_sorted.iter() {
-            let (vlabel, val) = (lit_vec[0].label(), lit_vec[0].polarity());
-            let mut bdd = self.var(vlabel, val);
-            for lit in lit_vec {
-                let (vlabel, val) = (lit.label(), lit.polarity());
-                let var = self.var(vlabel, val);
-                bdd = self.or(bdd, var);
-            }
-            cvec.push(bdd);
-        }
-        // now cvec has a list of all the clauses; collapse it down
-        let r = self.collapse_clauses(&cvec);
-        match r {
-            None => BddPtr::true_ptr(),
-            Some(x) => x,
-        }
-    }
-
     fn collapse_clauses(&'a self, vec: &[BddPtr<'a>]) -> Option<BddPtr<'a>> {
         if vec.is_empty() {
             None
@@ -155,76 +93,6 @@ pub trait BddBuilder<'a>: BottomUpBuilder<'a, BddPtr<'a>> {
                 (Some(v), None) | (None, Some(v)) => Some(v),
                 (Some(l), Some(r)) => Some(self.and(l, r)),
             }
-        }
-    }
-    fn compile_boolexpr(&'a self, expr: &LogicalExpr) -> BddPtr<'a> {
-        match &expr {
-            LogicalExpr::Literal(lbl, polarity) => self.var(VarLabel::new(*lbl as u64), *polarity),
-            LogicalExpr::And(ref l, ref r) => {
-                let r1 = self.compile_boolexpr(l);
-                let r2 = self.compile_boolexpr(r);
-                self.and(r1, r2)
-            }
-            LogicalExpr::Or(ref l, ref r) => {
-                let r1 = self.compile_boolexpr(l);
-                let r2 = self.compile_boolexpr(r);
-                self.or(r1, r2)
-            }
-            LogicalExpr::Not(ref e) => self.compile_boolexpr(e).neg(),
-            LogicalExpr::Iff(ref l, ref r) => {
-                let r1 = self.compile_boolexpr(l);
-                let r2 = self.compile_boolexpr(r);
-                self.iff(r1, r2)
-            }
-            LogicalExpr::Xor(ref l, ref r) => {
-                let r1 = self.compile_boolexpr(l);
-                let r2 = self.compile_boolexpr(r);
-                self.xor(r1, r2)
-            }
-            LogicalExpr::Ite {
-                ref guard,
-                ref thn,
-                ref els,
-            } => {
-                let g = self.compile_boolexpr(guard);
-                let t = self.compile_boolexpr(thn);
-                let e = self.compile_boolexpr(els);
-                self.ite(g, t, e)
-            }
-        }
-    }
-
-    /// Compiles a plan into a BDD
-    fn compile_plan(&'a self, expr: &BottomUpPlan) -> BddPtr<'a> {
-        match &expr {
-            BottomUpPlan::Literal(var, polarity) => self.var(*var, *polarity),
-            BottomUpPlan::And(ref l, ref r) => {
-                let r1 = self.compile_plan(l);
-                let r2 = self.compile_plan(r);
-                self.and(r1, r2)
-            }
-            BottomUpPlan::Or(ref l, ref r) => {
-                let r1 = self.compile_plan(l);
-                let r2 = self.compile_plan(r);
-                self.or(r1, r2)
-            }
-            BottomUpPlan::Iff(ref l, ref r) => {
-                let r1 = self.compile_plan(l);
-                let r2 = self.compile_plan(r);
-                self.iff(r1, r2)
-            }
-            BottomUpPlan::Ite(ref f, ref g, ref h) => {
-                let f = self.compile_plan(f);
-                let g = self.compile_plan(g);
-                let h = self.compile_plan(h);
-                self.ite(f, g, h)
-            }
-            BottomUpPlan::Not(ref f) => {
-                let f = self.compile_plan(f);
-                f.neg()
-            }
-            BottomUpPlan::ConstTrue => BddPtr::true_ptr(),
-            BottomUpPlan::ConstFalse => BddPtr::false_ptr(),
         }
     }
 }
@@ -316,5 +184,65 @@ where
         let a = self.and(iff, f);
 
         self.exists(a, lbl)
+    }
+
+    /// Compile a BDD from a CNF
+    fn compile_cnf(&'a self, cnf: &Cnf) -> BddPtr<'a> {
+        let mut cvec: Vec<BddPtr> = Vec::with_capacity(cnf.clauses().len());
+        if cnf.clauses().is_empty() {
+            return BddPtr::true_ptr();
+        }
+        // check if there is an empty clause -- if so, UNSAT
+        if cnf.clauses().iter().any(|x| x.is_empty()) {
+            return BddPtr::false_ptr();
+        }
+
+        // sort the clauses based on a best-effort bottom-up ordering of clauses
+        let mut cnf_sorted = cnf.clauses().to_vec();
+        cnf_sorted.sort_by(|c1, c2| {
+            // order the clause with the first-most variable last
+            let fst1 = c1
+                .iter()
+                .max_by(|l1, l2| {
+                    if self.less_than(l1.label(), l2.label()) {
+                        Ordering::Less
+                    } else {
+                        Ordering::Equal
+                    }
+                })
+                .unwrap();
+            let fst2 = c2
+                .iter()
+                .max_by(|l1, l2| {
+                    if self.less_than(l1.label(), l2.label()) {
+                        Ordering::Less
+                    } else {
+                        Ordering::Equal
+                    }
+                })
+                .unwrap();
+            if self.less_than(fst1.label(), fst2.label()) {
+                Ordering::Less
+            } else {
+                Ordering::Equal
+            }
+        });
+
+        for lit_vec in cnf_sorted.iter() {
+            let (vlabel, val) = (lit_vec[0].label(), lit_vec[0].polarity());
+            let mut bdd = self.var(vlabel, val);
+            for lit in lit_vec {
+                let (vlabel, val) = (lit.label(), lit.polarity());
+                let var = self.var(vlabel, val);
+                bdd = self.or(bdd, var);
+            }
+            cvec.push(bdd);
+        }
+        // now cvec has a list of all the clauses; collapse it down
+        let r = self.collapse_clauses(&cvec);
+        match r {
+            None => BddPtr::true_ptr(),
+            Some(x) => x,
+        }
     }
 }
